Rearrange test files a bit
