+++ bag_tracker/src/test/scala/uk/ac/wellcome/platform/archive/bag_tracker/AkkaBagTrackerClientTest.scala
@@ -1,8 +1,9 @@
-package uk.ac.wellcome.platform.archive.bag_tracker.client
+package uk.ac.wellcome.platform.archive.bag_tracker
 
 import akka.http.scaladsl.model.Uri
 import org.scalatest.concurrent.IntegrationPatience
 import uk.ac.wellcome.fixtures.TestWith
+import uk.ac.wellcome.platform.archive.bag_tracker.client.{AkkaBagTrackerClient, BagTrackerClient, BagTrackerClientTestCases}
 
 class AkkaBagTrackerClientTest extends BagTrackerClientTestCases with IntegrationPatience {
   override def withClient[R](testWith: TestWith[BagTrackerClient, R]): R =
