--- bag_tracker/src/test/scala/uk/ac/wellcome/platform/archive/bag_tracker/BagTrackerApiTest.scala
@@ -1,10 +1,0 @@
-package uk.ac.wellcome.platform.archive.bag_tracker
-
-import org.scalatest.funspec.AnyFunSpec
-import org.scalatest.matchers.should.Matchers
-
-class BagTrackerApiTest extends AnyFunSpec with Matchers {
-  it("asserts truth") {
-    true shouldBe true
-  }
-}
